Move CLI package to internal
